Better startup script

+++ stop.sh
@@ -1,3 +1,10 @@
 #!/bin/bash
 
-kill $( cat awkserver.pid )
+if [ -s awkserver.pid ]
+then
+    kill $( cat awkserver.pid )
+    rm awkserver.pid
+else
+    echo "awkserver.pid not found"
+    exit -1
+fi
